# public/swagger/vcaCommon.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "VCA",
        "contact": {},
        "version": "v1"
    },
    "servers": [
        {
            "url": "http://localhost:2173/vca/v1/api",
            "description": "Local"
        },
        {
            "url": "https://qa-api.tnrtp.org/idp-vca-service/vca/v1/api",
            "description": "QA"
        },
        {
            "url": "https://staging-api.tnrtp.org/idp-vca-service/vca/v1/api",
            "description": "Staging"
        },
        {
            "url": "https://dev-api.tnrtp.org/idp-vca-service/vca/v1/api",
            "description": "Development"
        },
        {
            "url": "https://prod-api.tnrtp.org/idp-vca-service/ydb/v1/api",
            "description": "Production"
        }
    ],
    "security": [
        {
            "Authorization": [],
            "token": []
        }
    ],
    "tags": [
        {
            "name": "VCA Report"
        }
    ],
    "paths": {
        "/vcaReport/createVcaReport": {
            "get": {
                "tags": ["VCA Report"],
                "summary": "Create VCA Report",
                "operationId": "CreateVCAReport",
                "parameters": [
                    {
                        "name": "activityId",
                        "in": "query",
                        "required": true,
                        "style": "form",
                        "explode": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "example": 1
                        }
                    },
                    {
                        "name": "commodityId",
                        "in": "query",
                        "required": false,
                        "style": "form",
                        "explode": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": { "description": "OK" }
                },
                "deprecated": false
            }
        },
        "/vcaReport/deleteVcaReport": {
            "delete": {
                "tags": ["VCA Report"],
                "summary": "Delete VCA Report",
                "operationId": "DeleteVCAReport",
                "parameters": [
                    {
                        "name": "reportId",
                        "in": "query",
                        "required": true,
                        "style": "form",
                        "explode": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": { "description": "OK" }
                },
                "deprecated": false
            }
        },
        "/vcaReport/hardDeleteVcaReport": {
            "delete": {
                "tags": ["VCA Report"],
                "summary": "Hard Delete VCA Report",
                "operationId": "HardDeleteVCAReport",
                "parameters": [],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/HardDeletedataRequest"
                            },
                            "example": {
                                "reportId": [1]
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": { "description": "OK" }
                },
                "deprecated": false
            }
        },
        "/vcaReport/submitVcaReport": {
            "get": {
                "tags": ["VCA Report"],
                "summary": "Submit VCA Report",
                "operationId": "SubmitVCAReport",
                "parameters": [
                    {
                        "name": "reportId",
                        "in": "query",
                        "required": true,
                        "style": "form",
                        "explode": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": { "description": "OK" }
                },
                "deprecated": false
            }
        },
        "/vcaReport/getVcaReportById": {
            "get": {
                "tags": ["VCA Report"],
                "summary": "Get VCA Report By ID Data",
                "operationId": "ViewReportData",
                "parameters": [
                    {
                        "name": "reportId",
                        "in": "query",
                        "required": true,
                        "style": "form",
                        "explode": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": { "description": "OK" }
                },
                "deprecated": false
            }
        },
        "/vcaReport/dashboard": {
            "get": {
                "tags": ["VCA Report"],
                "summary": "Get VCA Dashboard Data",
                "operationId": "VCADashboardData",
                "parameters": [
                    {
                        "name": "commodityId",
                        "in": "query",
                        "required": true,
                        "style": "form",
                        "explode": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": { "description": "OK" }
                },
                "deprecated": false
            }
        },
        "/vcaReport/reviewVcaReport": {
            "get": {
                "tags": ["VCA Report"],
                "summary": "Review VCA Report",
                "operationId": "ReviewVCAReport",
                "parameters": [
                    {
                        "name": "reportId",
                        "in": "query",
                        "required": true,
                        "style": "form",
                        "explode": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": { "description": "OK" }
                },
                "deprecated": false
            }
        },
        "/vcaReport/approveVcaReport": {
            "get": {
                "tags": ["VCA Report"],
                "summary": "Approve VCA Report",
                "operationId": "ApproveVCAReport",
                "parameters": [
                    {
                        "name": "reportId",
                        "in": "query",
                        "required": true,
                        "style": "form",
                        "explode": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": { "description": "OK" }
                },
                "deprecated": false
            }
        },
        "/vcaReport/getVcaReports": {
            "get": {
                "tags": ["VCA Report"],
                "summary": "Get VCA Reports",
                "operationId": "GetVCAReports",
                "parameters": [
                    {
                        "name": "districtId",
                        "in": "query",
                        "required": false,
                        "style": "form",
                        "explode": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "example": 2901
                        }
                    }
                ],
                "responses": {
                    "200": { "description": "OK" }
                },
                "deprecated": false
            }
        },
        "/vcaReport/getVcaReportLists": {
            "get": {
                "tags": ["VCA Report"],
                "summary": "Get VCA Report Lists",
                "operationId": "GetVCAReportLists",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "style": "form",
                        "explode": true,
                        "schema": {
                            "type": "string",
                            "example": ""
                        }
                    },
                    {
                        "name": "districtId",
                        "in": "query",
                        "required": false,
                        "style": "form",
                        "explode": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "example": 2901
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": true,
                        "style": "form",
                        "explode": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "example": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": true,
                        "style": "form",
                        "explode": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": { "description": "OK" }
                },
                "deprecated": false
            }
        },
        "/vcaReport/submitVcaComment": {
            "post": {
                "tags": ["VCA Report"],
                "summary": "Submit Comment",
                "operationId": "SubmitComment",
                "parameters": [],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SubmitCommentRequest"
                            },
                            "example": {
                                "reportId": 1,
                                "comment": "ss",
                                "replyCommentId": null,
                                "vcaCommentDocument": []
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": { "description": "OK" }
                },
                "deprecated": false
            }
        },
        "/vcaReport/getVcaComments": {
            "get": {
                "tags": ["VCA Report"],
                "summary": "Get Comments",
                "operationId": "GetComments",
                "parameters": [
                    {
                        "name": "reportId",
                        "in": "query",
                        "required": true,
                        "style": "form",
                        "explode": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "example": 13
                        }
                    },
                    {
                        "name": "showAll",
                        "in": "query",
                        "required": true,
                        "style": "form",
                        "explode": true,
                        "schema": {
                            "type": "boolean",
                            "example": false
                        }
                    }
                ],
                "responses": {
                    "200": { "description": "OK" }
                },
                "deprecated": false
            }
        },
        "/vcaReport/getVcaActivities": {
            "get": {
                "tags": ["VCA Report"],
                "summary": "Get Activities",
                "operationId": "GetActivities",
                "parameters": [
                    {
                        "name": "activityId",
                        "in": "query",
                        "required": true,
                        "style": "form",
                        "explode": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "example": 1
                        }
                    },
                    {
                        "name": "commodityId",
                        "in": "query",
                        "required": true,
                        "style": "form",
                        "explode": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": { "description": "OK" }
                },
                "deprecated": false
            }
        },
        "/vcaReport/getSingedUrl": {
            "post": {
                "tags": ["VCA Report"],
                "summary": "Get Signed Url",
                "operationId": "GetSignedUrldata",
                "parameters": [],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GetSignedUrldataRequest"
                            },
                            "example": {
                                "fileKeys": ["imagename"]
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": { "description": "OK" }
                },
                "deprecated": false
            }
        },
        "/vcaReport/uploadDoc": {
            "post": {
                "tags": ["VCA Report"],
                "summary": "Upload document",
                "operationId": "Uploaddocument",
                "parameters": [],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/UploadDocRequest"
                            },
                            "encoding": {}
                        }
                    },
                    "required": false
                },
                "responses": {
                    "200": { "description": "OK" }
                },
                "deprecated": false
            }
        }
    },
    "components": {
        "schemas": {
            "HardDeletedataRequest": {
                "title": "HardDeletedataRequest",
                "required": ["reportId"],
                "type": "object",
                "properties": {
                    "reportId": {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                },
                "example": {
                    "reportId": [1]
                }
            },
            "SubmitCommentRequest": {
                "title": "SubmitCommentRequest",
                "required": ["comment", "replyCommentId", "vcaCommentDocument", "reportId"],
                "type": "object",
                "properties": {
                    "reportId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "comment": {
                        "type": "string"
                    },
                    "replyCommentId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "vcaCommentDocument": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "example": {
                    "reportId": 1,
                    "comment": "ss",
                    "replyCommentId": 1,
                    "vcaCommentDocument": []
                }
            },
            "UploadDocRequest": {
                "required": ["file"],
                "type": "object",
                "properties": {
                    "file": {
                        "type": "string",
                        "format": "binary"
                    }
                }
            },
            "GetSignedUrldataRequest": {
                "title": "GetSignedUrldataRequest",
                "required": ["fileKeys"],
                "type": "object",
                "properties": {
                    "fileKeys": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "example": {
                    "fileKeys": ["imagename"]
                }
            }
        },
        "securitySchemes": {
            "Authorization": {
                "type": "apiKey",
                "in": "header",
                "name": "authorization"
            },
            "token": {
                "type": "apiKey",
                "in": "header",
                "name": "token"
            }
        }
    }
}